refactor(match): extract shared match helpers [GHI#101]
