+++ src/ui/web/modules/match/components/match-result/MatchResultOverlay.types.ts
@@ -1,9 +1,12 @@
-import type { UserAvatar } from "@/domain/entities/Avatar";
 import type {
   GameStatus,
   MatchScore,
   PlayerSlot,
 } from "@/domain/entities/Game";
+import type {
+  MatchResultParticipant,
+  MatchResultViewModel,
+} from "@/ui/shared/match/utils";
 
 export type ResultAccent = "primary" | "destructive";
 export type ResultIcon = "trophy" | "heart" | "wifi" | "flag";
@@ -19,32 +22,8 @@ export type MatchResultOverlayProps = {
   currentUserId?: string;
   score?: MatchScore;
   onPrimaryAction: () => Promise<void>;
-  currentUser: {
-    name: string;
-    avatar?: UserAvatar;
-  };
-  opponentUser: {
-    name: string;
-    avatar?: UserAvatar;
-  };
+  currentUser: MatchResultParticipant;
+  opponentUser: MatchResultParticipant;
 };
 
-export type MatchResultViewModel = {
-  title: string;
-  subtitle?: string;
-  accent: ResultAccent;
-  icon: ResultIcon;
-  pill?: string;
-  footer?: string;
-  score?: {
-    current: number;
-    opponent: number;
-  };
-  primaryLabel: string;
-  secondaryLabel: string;
-  changeModeLabel: string;
-  currentUser: MatchResultOverlayProps["currentUser"];
-  opponentUser: MatchResultOverlayProps["opponentUser"];
-  isCurrentWinner: boolean;
-  isAbandonedByCurrentUser?: boolean;
-};
+export type { MatchResultViewModel };
